\name{Rcmdrrmaeng2Plugin-internal}
\title{Internal Rcmdrrmaeng2Plugin objects}
\alias{centralLimitTheorem}
\alias{correlationExample}
\alias{flipCoin}
\alias{linearRegressionExample}
\alias{powerExample}
\alias{rollDie}
\alias{run.cor.examp}
\alias{simulateConfidenceIntervals}
\alias{slider}
\alias{visBinom}
\alias{visGamma}
\alias{visNormal}
\alias{vist}
\alias{mmultiWayAnova}
\alias{Menu.pb2level}
\alias{Menu.FrF2level}
\alias{Menu.fac}
\alias{Menu.oa}
\alias{Menu.ccd}
\alias{Menu.bbc}
\alias{Menu.lhs}
\description{Internal Rcmdrrmaeng2Plugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
