\name{Script Functions}
\alias{.add.frame}
\alias{.combine.rows}
\alias{.grid.matrix}
\alias{.initialize.grid.config}
\alias{.initialize.grid.rows}
\alias{.initialize.new.frames}
\alias{GUI_template}
\alias{anova_menu}


\title{Script Functions}

\description{List of internal functions used in the template scripts.}


\details{More information of the use of these functions can be found in the vignette located in the \code{doc} folder of the package.
  In here also the general script can be found in which these functions are used.}


\author{IMERMAENG2 \email{sf01363@naver.com}}
